requirement.txt
langchain
langchain-core
langchain-community
langchain-groq
chromadb
python-dotenv
beautifulsoup4
requests